refactored example_generator.py
