+++ libs/langchain/langchain/example_generator.py
@@ -1,23 +1,4 @@
-"""Utility functions for working with prompts."""
-from typing import List
+"""Keep here for backwards compatibility."""
+from langchain.utilities.example_generator import generate_example
 
-from langchain.chains.llm import LLMChain
-from langchain.prompts.few_shot import FewShotPromptTemplate
-from langchain.prompts.prompt import PromptTemplate
-from langchain.schema.language_model import BaseLanguageModel
-
-TEST_GEN_TEMPLATE_SUFFIX = "Add another example."
-
-
-def generate_example(
-    examples: List[dict], llm: BaseLanguageModel, prompt_template: PromptTemplate
-) -> str:
-    """Return another example given a list of examples for a prompt."""
-    prompt = FewShotPromptTemplate(
-        examples=examples,
-        suffix=TEST_GEN_TEMPLATE_SUFFIX,
-        input_variables=[],
-        example_prompt=prompt_template,
-    )
-    chain = LLMChain(llm=llm, prompt=prompt)
-    return chain.predict()
+__all__ = ["generate_example"]
